feat: make simpler backwards-compatible CourseRunCards and StatefulEnroll components
