+++ package.json
@@ -20,6 +20,8 @@
     "@fortawesome/free-regular-svg-icons": "5.15.1",
     "@fortawesome/free-solid-svg-icons": "5.15.1",
     "@fortawesome/react-fontawesome": "0.1.12",
+    "@tanstack/react-query": "^4.28.0",
+    "@tanstack/react-query-devtools": "^4.29.0",
     "algoliasearch": "4.6.0",
     "axios": "0.21.1",
     "classnames": "2.2.6",
