+++ src/components/app/App.jsx
@@ -3,6 +3,11 @@ import { Switch, Route } from 'react-router-dom';
 import { AppProvider, AuthenticatedPageRoute, PageRoute } from '@edx/frontend-platform/react';
 import { logError } from '@edx/frontend-platform/logging';
 import { initializeHotjar } from '@edx/frontend-enterprise-hotjar';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
 import AuthenticatedPage from './AuthenticatedPage';
 import EnterpriseAppPageRoutes from './EnterpriseAppPageRoutes';
@@ -15,6 +20,9 @@ import {
 import { EnterpriseInvitePage } from '../enterprise-invite';
 import { ExecutiveEducation2UPage } from '../executive-education-2u';
 import { ToastsProvider, Toasts } from '../Toasts';
+
+// Create a client
+const queryClient = new QueryClient();
 
 const App = () => {
   useEffect(() => {
@@ -32,29 +40,32 @@ const App = () => {
   }, []);
 
   return (
-    <AppProvider>
-      <NoticesProvider>
-        <ToastsProvider>
-          <Toasts />
-          <Switch>
-            <AuthenticatedPageRoute exact path="/" component={EnterpriseCustomerRedirect} />
-            <AuthenticatedPageRoute exact path="/r/:redirectPath+" component={EnterprisePageRedirect} />
-            <PageRoute exact path="/invite/:enterpriseCustomerInviteKey" component={EnterpriseInvitePage} />
-            <PageRoute
-              exact
-              path="/:enterpriseSlug/executive-education-2u"
-              render={(routeProps) => (
-                <AuthenticatedPage>
-                  <ExecutiveEducation2UPage {...routeProps} />
-                </AuthenticatedPage>
-              )}
-            />
-            <Route path="/:enterpriseSlug" component={EnterpriseAppPageRoutes} />
-            <PageRoute path="*" component={NotFoundPage} />
-          </Switch>
-        </ToastsProvider>
-      </NoticesProvider>
-    </AppProvider>
+    <QueryClientProvider client={queryClient}>
+      <ReactQueryDevtools initialIsOpen={false} />
+      <AppProvider>
+        <NoticesProvider>
+          <ToastsProvider>
+            <Toasts />
+            <Switch>
+              <AuthenticatedPageRoute exact path="/" component={EnterpriseCustomerRedirect} />
+              <AuthenticatedPageRoute exact path="/r/:redirectPath+" component={EnterprisePageRedirect} />
+              <PageRoute exact path="/invite/:enterpriseCustomerInviteKey" component={EnterpriseInvitePage} />
+              <PageRoute
+                exact
+                path="/:enterpriseSlug/executive-education-2u"
+                render={(routeProps) => (
+                  <AuthenticatedPage>
+                    <ExecutiveEducation2UPage {...routeProps} />
+                  </AuthenticatedPage>
+                )}
+              />
+              <Route path="/:enterpriseSlug" component={EnterpriseAppPageRoutes} />
+              <PageRoute path="*" component={NotFoundPage} />
+            </Switch>
+          </ToastsProvider>
+        </NoticesProvider>
+      </AppProvider>
+    </QueryClientProvider>
   );
 };
