+++ src/components/stateful-enroll/StatefulEnroll.jsx
@@ -1,8 +1,19 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { StatefulButton } from '@edx/paragon';
+import { getAuthenticatedUser } from '@edx/frontend-platform/auth';
 
-const StatefulEnroll = () => {
+import {
+  useRedemptionMutation,
+  useTransactionStatus,
+} from './data';
+
+const StatefulEnroll = ({
+  contentKey,
+}) => {
   const [enrollButtonState, setEnrollButtonState] = useState('default');
+  const [transactionUUID, setTransactionUUID] = useState();
+
   const enrollButtonLabels = {
     default: 'Enroll',
     pending: 'Enrolling...',
@@ -10,7 +21,37 @@ const StatefulEnroll = () => {
     error: 'Try again',
   };
 
-  const handleEnrollButtonClick = () => {};
+  const redemptionMutation = useRedemptionMutation({
+    onMutate: () => {
+      setEnrollButtonState('pending');
+    },
+    onSuccess: (transaction) => {
+      setTransactionUUID(transaction.uuid);
+    },
+    onError: () => {
+      setEnrollButtonState('error');
+    },
+  });
+
+  useTransactionStatus({
+    transactionUUID,
+    onSuccess: (transaction) => {
+      setEnrollButtonState('complete');
+      const coursewareURL = transaction.coursewareRedirectUrl;
+      console.log(`[EMET] Successfully enrolled. Redirecting to courseware URL (${coursewareURL})!`);
+      // window.location.href = coursewareURL;
+      setTimeout(() => {
+        setEnrollButtonState('default');
+      }, 2000);
+    },
+  });
+
+  const handleEnrollButtonClick = () => {
+    redemptionMutation.mutate({
+      userId: getAuthenticatedUser().id,
+      contentKey,
+    });
+  };
 
   return (
     <StatefulButton
@@ -19,6 +60,10 @@ const StatefulEnroll = () => {
       onClick={handleEnrollButtonClick}
     />
   );
+};
+
+StatefulEnroll.propTypes = {
+  contentKey: PropTypes.string.isRequired,
 };
 
 export default StatefulEnroll;
