+++ src/components/stateful-enroll/index.js
@@ -1,0 +1,5 @@
+import StatefulEnroll from './StatefulEnroll';
+
+export * from './data';
+
+export default StatefulEnroll;
